# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/gearboxes/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Gearbox",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Gearboxes"
                ],
                "summary": "Create a Gearbox",
                "parameters": [
                    {
                        "description": "Create a Gearbox",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.CreateGearboxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Gearbox response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.GearboxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/gearboxes/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Gearboxes",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Gearboxes"
                ],
                "summary": "Get Gearboxes",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gearbox response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.PagedList-github_com_salmantaghooni_golang-car-web-api_src_api_dto_GearboxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/gearboxes/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a Gearbox",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Gearboxes"
                ],
                "summary": "Get a Gearbox",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gearbox response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.GearboxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a Gearbox",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Gearboxes"
                ],
                "summary": "Update a Gearbox",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a Gearbox",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.UpdateGearboxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gearbox response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.GearboxResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a Gearbox",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Gearboxes"
                ],
                "summary": "Delete a Gearbox",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.CreateGearboxRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.GearboxResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.PagedList-github_com_salmantaghooni_golang-car-web-api_src_api_dto_GearboxResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.GearboxResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_dto.Sort"
                    }
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_dto.UpdateGearboxRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "github_com_salmantaghooni_golang-car-web-api_src_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_salmantaghooni_golang-car-web-api_src_api_validations.ValidationError"
                    }
                }
            }
        },
        "github_com_salmantaghooni_golang-car-web-api_src_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_salmantaghooni_golang-car-web-api_src_api_validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "peroperty": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}